Add rejected edges back to the output with undefined value
When running d3.ForceBundle some of the edges are filtered out but I don't know which edges so I no longer could make a link between the original data set and the results output from d3.ForceBundle. That is why I added those filtered edges back to the d3.ForceBundle output with the undefined value.